# Conflicts:
#	src/main/java/com/action/SupplierAction.java
2019.06.14

+++ src/main/java/com/service/sup/SupplierTrademarkServiceImpl.java
@@ -1,6 +1,7 @@
 package com.service.sup;
 
 import com.beans.SupplierTrademark;
+import com.dao.sup.MiddleMapper;
 import com.dao.sup.SupplierTrademarkMapper;
 import com.util.Page;
 import org.springframework.stereotype.Service;
@@ -21,6 +22,8 @@ public class SupplierTrademarkServiceImpl implements SupplierTrademarkService {
 
     @Resource
     private SupplierTrademarkMapper supplierTrademarkMapper;
+    @Resource
+    private MiddleMapper middleMapper;
     //添加品牌信息
     @Override
     public int addSupplierTrademark(SupplierTrademark supplierTrademark) {
@@ -29,7 +32,13 @@ public class SupplierTrademarkServiceImpl implements SupplierTrademarkService {
     //删除品牌信息
     @Override
     public int deleteTrademark(int id) {
-        return supplierTrademarkMapper.deleteSrademark(id);
+            int  a =middleMapper.queryCountbyStaff(id);
+            if (a>0){
+                return 0;
+            }else{
+                 return supplierTrademarkMapper.deleteSrademark(id);
+            }
+
     }
     //更改品牌信息
     @Override
@@ -45,7 +54,7 @@ public class SupplierTrademarkServiceImpl implements SupplierTrademarkService {
             if (pageIndex == 0) {
                 pageIndex = 1;
             }
-            page.setPageSize(2);
+            page.setPageSize(10);
             page.setTotalCount(supplierTrademarkMapper.querycount(name,product,enterpriseName));
             page.setCurrentPageNo(pageIndex);
             List<SupplierTrademark> list= supplierTrademarkMapper.querybysome(name,product,enterpriseName,(page.getCurrentPageNo()-1)*page.getPageSize(),page.getPageSize());
